+++ src/duke/Duke.java
@@ -1,4 +1,15 @@
-package main.java;
+package duke;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
+
+import duke.util.DukeException;
+import duke.util.Parser;
+import duke.util.Storage;
+import duke.util.TaskList;
+import duke.util.Ui;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
