+++ src/duke/task/Event.java
@@ -1,4 +1,4 @@
-package main.java;
+package duke.task;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
