+++ src/duke/util/DukeException.java
@@ -1,4 +1,4 @@
-package main.java;
+package duke.util;
 
 public class DukeException extends Exception {
     public DukeException(String msg) {
