+++ src/duke/util/Storage.java
@@ -1,4 +1,9 @@
-package main.java;
+package duke.util;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
 
 import java.io.*;
 import java.time.format.DateTimeFormatter;
