+++ src/duke/util/TaskList.java
@@ -1,4 +1,6 @@
-package main.java;
+package duke.util;
+
+import duke.task.Task;
 
 import java.util.ArrayList;
 import java.util.List;
